// services/products/types/Product.type.ts
import { ItemType } from "@/types/Item.type";

export type ProductType = {
  id: string;
  name: string;
  slug: string;
  description: string;
  shortDescription: string;
  sku: string;
  price: string;
  comparePrice: string;
  costPrice: string;
  stock: number;
  lowStockThreshold: number;
  weight: string;
  dimensions: {
    length: number;
    width: number;
    height: number;
  };
  images: string[];
  thumbnail: string;
  categoryId: string;
  vendorId: null;
  brandId: string;
  tags: string[];
  isActive: boolean;
  isFeatured: boolean;
  isDigital: boolean;
  downloadUrl: null | string;
  taxable: boolean;
  taxRate: string;
  rating: string;
  reviewCount: number;
  viewCount: number;
  soldCount: number;
  publishedAt: null;
  createdAt: string;
  updatedAt: string;
  category: ItemType;
  discountPercentage: number;
  inStock: boolean;
  isLowStock: boolean;
  metaTitle: string;
  metaDescription: string;
  locale: string;
};